add dry run

+++ scripts/binary_release.sh
@@ -23,6 +23,7 @@ fi
 COMMIT=$(git rev-parse HEAD)
 
 RELEASE_TAG=""
+# the release commit is assigned multiple tags. Only use the tag from the git run.
 if [ "${GITHUB_EVENT_NAME}" = "release" ] && [ -n "${GITHUB_REF_NAME}" ]; then
   RELEASE_TAG="${GITHUB_REF_NAME}"
 fi
@@ -33,19 +34,35 @@ os_and_arch=${os}-${arch}
 
 # cspell:words czvf
 function tar_binary {
+  if [ "${DRY_RUN}" = "true" ]; then
+    echo "DRY RUN MODE: Would run 'tar -C $1 -czvf $1.tar.gz ${os_and_arch}'"
+    touch $1.tar.gz
+    return
+  fi
   mkdir -p $1/${os_and_arch}
   cp target/release/$1 $1/${os_and_arch}/
   tar -C $1 -czvf $1.tar.gz ${os_and_arch}
 }
 
-make $release
+if [ "${DRY_RUN}" = "true" ]; then
+  echo "DRY RUN MODE: Would run 'make $release'"
+else
+  make $release
+fi
 
 function upload_binary {
+  DRY_RUN_FLAG=""
+  if [ "${DRY_RUN}" = "true" ]; then
+    DRY_RUN_FLAG="--dryrun"
+  fi
+
   if [ "$release" = "release" ]
   then
     tar_binary $1
     tar_file=$1.tar.gz
 
+    # If it is a release, we will upload to the tag as well.
+    # Maybe we do not need to upload to the branch anymore.
     upload_targets=()
     if [ -n "${BRANCH}" ]; then
       upload_targets+=("${BRANCH}")
@@ -59,19 +76,22 @@ function upload_binary {
       exit 1
     fi
 
-    for target in "${upload_targets[@]}"; do
-      aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os_and_arch}/${target}/$1
-      aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os_and_arch}/${target}/${COMMIT}/$1
-      aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os_and_arch}/${target}/${COMMIT}/stable/$1
-
-      aws s3 cp --acl public-read ${tar_file} s3://build.nearprotocol.com/nearcore/${os_and_arch}/${target}/${tar_file}
-      aws s3 cp --acl public-read ${tar_file} s3://build.nearprotocol.com/nearcore/${os_and_arch}/${target}/${COMMIT}/${tar_file}
-      aws s3 cp --acl public-read ${tar_file} s3://build.nearprotocol.com/nearcore/${os_and_arch}/${target}/${COMMIT}/stable/${tar_file}
+    sources=("target/release/$1" "${tar_file}")
+    destinations=("$1" "${tar_file}")
+    
+    for target in "${upload_targets[@]}"; do  
+      for i in "${!sources[@]}"; do
+        src=${sources[$i]}
+        dst=${destinations[$i]}
+        aws s3 cp --acl public-read ${src} s3://build.nearprotocol.com/nearcore/${os_and_arch}/${target}/${dst} ${DRY_RUN_FLAG}
+        aws s3 cp --acl public-read ${src} s3://build.nearprotocol.com/nearcore/${os_and_arch}/${target}/${COMMIT}/${dst} ${DRY_RUN_FLAG}
+        aws s3 cp --acl public-read ${src} s3://build.nearprotocol.com/nearcore/${os_and_arch}/${target}/${COMMIT}/stable/${dst} ${DRY_RUN_FLAG}
+      done
     done
 
   else
     folder="${release%-release}"
-    aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/${COMMIT}/${folder}/$1
+    aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/${COMMIT}/${folder}/$1 ${DRY_RUN_FLAG}
   fi
 }
